// src/navigators-types.ts
export type NavigatorKind = 'OnboarderNavigator' | 'ERC20TributeNavigator' | 'NFTGatedNavigator' | 'SignalNavigator' | 'TimelockNavigator' | 'VestingNavigator' | 'BudgetNavigator' | 'SubscriptionNavigator';
export interface NavigatorDeployConfig {
  OnboarderNavigator: {
    daoShip: string;
    shareMultiplier: bigint;
    lootMultiplier: bigint;
    pricePerUnit: bigint;
    sharesPerUnit: bigint;
    lootPerUnit: bigint;
    minTribute: bigint;
    expiry: bigint;
    mintCap: bigint;
    perAddressCap: bigint;
    allowlistRoot: string;
    name: string;
    description: string;
  };
  ERC20TributeNavigator: {
    daoShip: string;
    tributeToken: string;
    pricePerShare: bigint;
    pricePerLoot: bigint;
    expiry: bigint;
    mintCap: bigint;
    perAddressCap: bigint;
    allowlistRoot: string;
    name: string;
    description: string;
  };
  NFTGatedNavigator: {
    daoShip: string;
    gateToken: string;
    sharesPerHolder: bigint;
    lootPerHolder: bigint;
    requireTribute: boolean;
    tributeAmount: bigint;
    expiry: bigint;
    mintCap: bigint;
    perAddressCap: bigint;
    allowlistRoot: string;
    name: string;
    description: string;
  };
  SignalNavigator: {
    daoShip: string;
    minSharesToCreatePoll: bigint;
    minDuration: bigint;
    maxDuration: bigint;
    maxStartDelay: bigint;
    name: string;
    description: string;
  };
  TimelockNavigator: {
    daoShip: string;
    delay: bigint;
    expiryWindow: bigint;
    name: string;
    description: string;
  };
  VestingNavigator: {
    daoShip: string;
    name: string;
    description: string;
  };
  BudgetNavigator: {
    daoShip: string;
    name: string;
    description: string;
  };
  SubscriptionNavigator: {
    daoShip: string;
    tokens: readonly string[];
    feesPerPeriod: readonly bigint[];
    periodDuration: bigint;
    graceDuration: bigint;
    startTime: bigint;
    collectorRewardBps: bigint;
    burnOnCollect: boolean;
    initialMembers: readonly string[];
    name: string;
    description: string;
  };
}
export interface NavigatorReads {
  OnboarderNavigator: {
    'allowlistRoot()': readonly [];
    'allowlistRoot': readonly [];
    'daoShip()': readonly [];
    'daoShip': readonly [];
    'deployer()': readonly [];
    'deployer': readonly [];
    'expiry()': readonly [];
    'expiry': readonly [];
    'lootMultiplier()': readonly [];
    'lootMultiplier': readonly [];
    'lootPerUnit()': readonly [];
    'lootPerUnit': readonly [];
    'minTribute()': readonly [];
    'minTribute': readonly [];
    'mintCap()': readonly [];
    'mintCap': readonly [];
    'mintedTo(address)': readonly [arg0: string];
    'mintedTo': readonly [arg0: string];
    'navigatorType()': readonly [];
    'navigatorType': readonly [];
    'paused()': readonly [];
    'paused': readonly [];
    'perAddressCap()': readonly [];
    'perAddressCap': readonly [];
    'pricePerUnit()': readonly [];
    'pricePerUnit': readonly [];
    'shareMultiplier()': readonly [];
    'shareMultiplier': readonly [];
    'sharesPerUnit()': readonly [];
    'sharesPerUnit': readonly [];
    'totalMinted()': readonly [];
    'totalMinted': readonly [];
  };
  ERC20TributeNavigator: {
    'allowlistRoot()': readonly [];
    'allowlistRoot': readonly [];
    'daoShip()': readonly [];
    'daoShip': readonly [];
    'deployer()': readonly [];
    'deployer': readonly [];
    'expiry()': readonly [];
    'expiry': readonly [];
    'mintCap()': readonly [];
    'mintCap': readonly [];
    'mintedTo(address)': readonly [arg0: string];
    'mintedTo': readonly [arg0: string];
    'navigatorType()': readonly [];
    'navigatorType': readonly [];
    'paused()': readonly [];
    'paused': readonly [];
    'perAddressCap()': readonly [];
    'perAddressCap': readonly [];
    'pricePerLoot()': readonly [];
    'pricePerLoot': readonly [];
    'pricePerShare()': readonly [];
    'pricePerShare': readonly [];
    'totalMinted()': readonly [];
    'totalMinted': readonly [];
    'tributeToken()': readonly [];
    'tributeToken': readonly [];
  };
  NFTGatedNavigator: {
    'allowlistRoot()': readonly [];
    'allowlistRoot': readonly [];
    'canOnboard(address,uint256,bytes32[])': readonly [candidate: string, tokenId: bigint, proof: readonly string[]];
    'canOnboard(address,uint256)': readonly [candidate: string, tokenId: bigint];
    'claimed(uint256)': readonly [arg0: bigint];
    'claimed': readonly [arg0: bigint];
    'daoShip()': readonly [];
    'daoShip': readonly [];
    'deployer()': readonly [];
    'deployer': readonly [];
    'expiry()': readonly [];
    'expiry': readonly [];
    'gateToken()': readonly [];
    'gateToken': readonly [];
    'isEligible(address)': readonly [candidate: string];
    'isEligible': readonly [candidate: string];
    'isEligibleToken(address,uint256)': readonly [candidate: string, tokenId: bigint];
    'isEligibleToken': readonly [candidate: string, tokenId: bigint];
    'lootPerHolder()': readonly [];
    'lootPerHolder': readonly [];
    'mintCap()': readonly [];
    'mintCap': readonly [];
    'mintedTo(address)': readonly [arg0: string];
    'mintedTo': readonly [arg0: string];
    'navigatorType()': readonly [];
    'navigatorType': readonly [];
    'paused()': readonly [];
    'paused': readonly [];
    'perAddressCap()': readonly [];
    'perAddressCap': readonly [];
    'requireTribute()': readonly [];
    'requireTribute': readonly [];
    'sharesPerHolder()': readonly [];
    'sharesPerHolder': readonly [];
    'totalMinted()': readonly [];
    'totalMinted': readonly [];
    'tributeAmount()': readonly [];
    'tributeAmount': readonly [];
  };
  SignalNavigator: {
    'MAX_OPTIONS()': readonly [];
    'MAX_OPTIONS': readonly [];
    'MAX_WINDOW()': readonly [];
    'MAX_WINDOW': readonly [];
    'MIN_OPTIONS()': readonly [];
    'MIN_OPTIONS': readonly [];
    'daoShip()': readonly [];
    'daoShip': readonly [];
    'deployer()': readonly [];
    'deployer': readonly [];
    'getOptionVotes(uint256,uint8)': readonly [pollId: bigint, option: bigint];
    'getOptionVotes': readonly [pollId: bigint, option: bigint];
    'getResults(uint256)': readonly [pollId: bigint];
    'getResults': readonly [pollId: bigint];
    'hasVoted(uint256,address)': readonly [pollId: bigint, voter: string];
    'hasVoted': readonly [pollId: bigint, voter: string];
    'maxDuration()': readonly [];
    'maxDuration': readonly [];
    'maxStartDelay()': readonly [];
    'maxStartDelay': readonly [];
    'minDuration()': readonly [];
    'minDuration': readonly [];
    'minSharesToCreatePoll()': readonly [];
    'minSharesToCreatePoll': readonly [];
    'navigatorType()': readonly [];
    'navigatorType': readonly [];
    'pollCount()': readonly [];
    'pollCount': readonly [];
    'pollStatus(uint256)': readonly [pollId: bigint];
    'pollStatus': readonly [pollId: bigint];
    'polls(uint256)': readonly [arg0: bigint];
    'polls': readonly [arg0: bigint];
  };
  TimelockNavigator: {
    'MAX_DELAY()': readonly [];
    'MAX_DELAY': readonly [];
    'MAX_EXPIRY()': readonly [];
    'MAX_EXPIRY': readonly [];
    'MIN_DELAY()': readonly [];
    'MIN_DELAY': readonly [];
    'MIN_EXPIRY()': readonly [];
    'MIN_EXPIRY': readonly [];
    'RECOMMENDED_DELAY()': readonly [];
    'RECOMMENDED_DELAY': readonly [];
    'changeCount()': readonly [];
    'changeCount': readonly [];
    'daoShip()': readonly [];
    'daoShip': readonly [];
    'delay()': readonly [];
    'delay': readonly [];
    'deployer()': readonly [];
    'deployer': readonly [];
    'expiryWindow()': readonly [];
    'expiryWindow': readonly [];
    'isExecutable(uint256)': readonly [changeId: bigint];
    'isExecutable': readonly [changeId: bigint];
    'navigatorType()': readonly [];
    'navigatorType': readonly [];
    'paused()': readonly [];
    'paused': readonly [];
    'queuedChanges(uint256)': readonly [arg0: bigint];
    'queuedChanges': readonly [arg0: bigint];
  };
// ...
  SubscriptionNavigator: {
    'MAX_COLLECTOR_BPS()': readonly [];
    'MAX_COLLECTOR_BPS': readonly [];
    'MAX_PERIOD()': readonly [];
    'MAX_PERIOD': readonly [];
    'MIN_PERIOD()': readonly [];
    'MIN_PERIOD': readonly [];
    'NATIVE()': readonly [];
    'NATIVE': readonly [];
    'acceptedTokenCount()': readonly [];
    'acceptedTokenCount': readonly [];
    'acceptedTokens(uint256)': readonly [arg0: bigint];
    'acceptedTokens': readonly [arg0: bigint];
    'burnOnCollect()': readonly [];
    'burnOnCollect': readonly [];
    'collectorRewardBps()': readonly [];
    'collectorRewardBps': readonly [];
    'daoShip()': readonly [];
    'daoShip': readonly [];
    'deployer()': readonly [];
    'deployer': readonly [];
    'feePerPeriod(address)': readonly [arg0: string];
    'feePerPeriod': readonly [arg0: string];
    'getAcceptedTokens()': readonly [];
    'getAcceptedTokens': readonly [];
    'graceDuration()': readonly [];
    'graceDuration': readonly [];
    'inGracePeriod(address)': readonly [member: string];
    'inGracePeriod': readonly [member: string];
    'isCurrent(address)': readonly [member: string];
    'isCurrent': readonly [member: string];
    'isDelinquent(address)': readonly [member: string];
    'isDelinquent': readonly [member: string];
    'isEnrolled(address)': readonly [member: string];
    'isEnrolled': readonly [member: string];
    'navigatorType()': readonly [];
    'navigatorType': readonly [];
    'nextDeadline(address)': readonly [member: string];
    'nextDeadline': readonly [member: string];
    'paidThrough(address)': readonly [arg0: string];
    'paidThrough': readonly [arg0: string];
    'paused()': readonly [];
    'paused': readonly [];
    'periodDuration()': readonly [];
    'periodDuration': readonly [];
    'quote(uint256,address)': readonly [periods: bigint, token: string];
    'quote': readonly [periods: bigint, token: string];
    'startTime()': readonly [];
    'startTime': readonly [];
  };
}
export interface NavigatorWrites {
  OnboarderNavigator: {
    'onboard(bytes32[])': readonly [proof: readonly string[]];
    'onboard()': readonly [];
    'pause()': readonly [];
    'pause': readonly [];
    'unpause()': readonly [];
    'unpause': readonly [];
    'withdrawStuckETH(address,uint256)': readonly [to: string, amount: bigint];
    'withdrawStuckETH': readonly [to: string, amount: bigint];
  };
// ...
}
export interface NavigatorReadResults {
  OnboarderNavigator: {
    'allowlistRoot()': string;
    'allowlistRoot': string;
    'daoShip()': string;
    'daoShip': string;
    'deployer()': string;
    'deployer': string;
    'expiry()': bigint;
    'expiry': bigint;
    'lootMultiplier()': bigint;
    'lootMultiplier': bigint;
    'lootPerUnit()': bigint;
    'lootPerUnit': bigint;
    'minTribute()': bigint;
    'minTribute': bigint;
    'mintCap()': bigint;
    'mintCap': bigint;
    'mintedTo(address)': bigint;
    'mintedTo': bigint;
    'navigatorType()': string;
    'navigatorType': string;
    'paused()': boolean;
    'paused': boolean;
    'perAddressCap()': bigint;
    'perAddressCap': bigint;
    'pricePerUnit()': bigint;
    'pricePerUnit': bigint;
    'shareMultiplier()': bigint;
    'shareMultiplier': bigint;
    'sharesPerUnit()': bigint;
    'sharesPerUnit': bigint;
    'totalMinted()': bigint;
    'totalMinted': bigint;
  };
  ERC20TributeNavigator: {
    'allowlistRoot()': string;
    'allowlistRoot': string;
    'daoShip()': string;
    'daoShip': string;
    'deployer()': string;
    'deployer': string;
    'expiry()': bigint;
    'expiry': bigint;
    'mintCap()': bigint;
    'mintCap': bigint;
    'mintedTo(address)': bigint;
    'mintedTo': bigint;
    'navigatorType()': string;
    'navigatorType': string;
    'paused()': boolean;
    'paused': boolean;
    'perAddressCap()': bigint;
    'perAddressCap': bigint;
    'pricePerLoot()': bigint;
    'pricePerLoot': bigint;
    'pricePerShare()': bigint;
    'pricePerShare': bigint;
    'totalMinted()': bigint;
    'totalMinted': bigint;
    'tributeToken()': string;
    'tributeToken': string;
  };
// ...
}